service_management/requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
djangorestframework==3.10.3
html5lib==1.0.1
mysqlclient==1.3.12
Pillow==6.1.0
PyPDF2==1.26.0
pytz==2019.2
reportlab==3.5.28
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
